DM: Update the vcpu id type as uint16_t for vm_create_vcpu
Update the vcpu id type as uint16_t for vm_create_vcpu, this
keeps alignment with the updates for the structure acrn_create_vcpu
used by hcall_create_vcpu in the hypervisor.

In the device model, the caller is responsible for vcpu id type
conversion; vcpu id type is uint16_t for external interface in the
current implement.

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ devicemodel/core/vmmapi.c
@@ -707,7 +707,7 @@ vm_reset_ptdev_intx_info(struct vmctx *ctx, int virt_pin, bool pic_pin)
 }
 
 int
-vm_create_vcpu(struct vmctx *ctx, int vcpu_id)
+vm_create_vcpu(struct vmctx *ctx, uint16_t vcpu_id)
 {
 	struct acrn_create_vcpu cv;
 	int error;
